1238: rename more CaseExternal related files
