+++ shared/src/business/entities/cases/ElectronicPetitionForMinorWithoutGuardianContacts.test.ts
@@ -1,17 +1,17 @@
 import { CASE_TYPES_MAP, COUNTRY_TYPES, PARTY_TYPES } from '../EntityConstants';
-import { CaseExternal } from './ElectronicPetition';
+import { ElectronicPetition } from './ElectronicPetition';
 import { applicationContext } from '../../test/createTestApplicationContext';
 
 describe('CaseExternal', () => {
-  describe('for Partnership (as the Tax Matters Partner) Contacts', () => {
+  describe('for Minor without Guardian Contacts', () => {
     it('should not validate without contacts', () => {
-      const caseExternal = new CaseExternal(
+      const caseExternal = new ElectronicPetition(
         {
           caseType: CASE_TYPES_MAP.other,
           filingType: 'Myself',
           hasIrsNotice: true,
           irsNoticeDate: '2009-10-13',
-          partyType: PARTY_TYPES.partnershipAsTaxMattersPartner,
+          partyType: PARTY_TYPES.nextFriendForMinor,
           petitionFile: {},
           petitionFileSize: 1,
           preferredTrialCity: 'Memphis, Tennessee',
@@ -26,7 +26,7 @@ describe('CaseExternal', () => {
     });
 
     it('can validate contacts', () => {
-      const caseExternal = new CaseExternal(
+      const caseExternal = new ElectronicPetition(
         {
           caseType: CASE_TYPES_MAP.other,
           contactPrimary: {
@@ -44,7 +44,7 @@ describe('CaseExternal', () => {
           filingType: 'Myself',
           hasIrsNotice: true,
           irsNoticeDate: '2009-10-13',
-          partyType: PARTY_TYPES.partnershipAsTaxMattersPartner,
+          partyType: PARTY_TYPES.nextFriendForMinor,
           petitionFile: {},
           petitionFileSize: 1,
           preferredTrialCity: 'Memphis, Tennessee',
