+++ shared/src/business/entities/cases/ElectronicPetitionForPetitionerAndSpouseContacts.test.ts
@@ -1,17 +1,17 @@
 import { CASE_TYPES_MAP, COUNTRY_TYPES, PARTY_TYPES } from '../EntityConstants';
-import { CaseExternal } from './ElectronicPetition';
+import { ElectronicPetition } from './ElectronicPetition';
 import { applicationContext } from '../../test/createTestApplicationContext';
 
 describe('CaseExternal', () => {
-  describe('for Corporation Contacts', () => {
-    it('should not validate without contact', () => {
-      const caseExternal = new CaseExternal(
+  describe('for Petitioner And Spouse Contacts', () => {
+    it('should not validate without contacts', () => {
+      const caseExternal = new ElectronicPetition(
         {
           caseType: CASE_TYPES_MAP.other,
           filingType: 'Myself',
           hasIrsNotice: true,
           irsNoticeDate: '2009-10-13',
-          partyType: PARTY_TYPES.corporation,
+          partyType: PARTY_TYPES.petitionerSpouse,
           petitionFile: {},
           petitionFileSize: 1,
           preferredTrialCity: 'Memphis, Tennessee',
@@ -25,8 +25,8 @@ describe('CaseExternal', () => {
       expect(caseExternal.isValid()).toEqual(false);
     });
 
-    it('should not validate without inCareOf', () => {
-      const caseExternal = new CaseExternal(
+    it('can validate primary contact name', () => {
+      const caseExternal = new ElectronicPetition(
         {
           caseType: CASE_TYPES_MAP.other,
           contactPrimary: {
@@ -40,41 +40,20 @@ describe('CaseExternal', () => {
             postalCode: '05198',
             state: 'AK',
           },
-          filingType: 'Myself',
-          hasIrsNotice: true,
-          irsNoticeDate: '2009-10-13',
-          partyType: PARTY_TYPES.corporation,
-          petitionFile: {},
-          petitionFileSize: 1,
-          preferredTrialCity: 'Memphis, Tennessee',
-          procedureType: 'Small',
-          signature: true,
-        },
-        { applicationContext },
-      );
-      expect(caseExternal.isValid()).toEqual(false);
-    });
-
-    it('can validate primary contact', () => {
-      const caseExternal = new CaseExternal(
-        {
-          caseType: CASE_TYPES_MAP.other,
-          contactPrimary: {
-            address1: '876 12th Ave',
-            city: 'Nashville',
-            country: 'USA',
+          contactSecondary: {
+            address1: '1599 Pennsylvania Ave',
+            city: 'Walla Walla',
             countryType: COUNTRY_TYPES.DOMESTIC,
             email: 'someone@example.com',
-            inCareOf: 'USTC',
-            name: 'Jimmy Dean',
+            name: 'Betty Crocker',
             phone: '1234567890',
-            postalCode: '05198',
-            state: 'AK',
+            postalCode: '78774',
+            state: 'WA',
           },
           filingType: 'Myself',
           hasIrsNotice: true,
           irsNoticeDate: '2009-10-13',
-          partyType: PARTY_TYPES.corporation,
+          partyType: PARTY_TYPES.petitionerSpouse,
           petitionFile: {},
           petitionFileSize: 1,
           preferredTrialCity: 'Memphis, Tennessee',
